add support for primitive sizes with vertices > 65535
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1377 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ primitive.h
@@ -12,7 +12,7 @@ class Mesh
 {
 public:
     std::vector<Vertex3D_NoTex2> m_vertices;
-    std::vector<WORD> m_indices;
+    std::vector<unsigned int> m_indices;
 
     void Clear();
     bool LoadWavefrontObj(const char *fname, bool flipTV, bool convertToLeftHanded);
